01_Local_Engine/requirements.txt
requests
pandas
numpy
supabase
python-dotenv
tqdm
openpyxl
PyPDF2
